requirements.txt
# Core
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# AI Services
anthropic==0.30.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
Pillow==10.3.0

# Async Tasks
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Email
django-smtp-ssl==1.0

# Database
dj-database-url==2.2.0

# File Storage
boto3==1.34.131  # Para S3 si lo necesitan después
django-storages==1.14.3

# Utilities
python-dateutil==2.9.0
pytz==2024.1
pandas==2.2.2
openpyxl==3.1.4  # Para exportar reportes Excel

# WebSockets (para dashboard en tiempo real)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Development
django-extensions==3.2.3
ipython==8.26.0

# Testing
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0
faker==25.8.0
django-filter==24.2
